fix splitting by files
