+++ react-ts-template/src/api/charactersPost.ts
@@ -1,5 +1,5 @@
 import instance from "./helpers/axios";
-import { Characters, Comics, Character, Comic } from './types/post';
+import { Characters, Character } from './types/characters';
 import md5 from 'md5';
 
 const ts = Date.now();
@@ -32,33 +32,7 @@ export default {
           return response.data.data.results[0];
   },
 
-    async getComicsList(offset: number): Promise<{ results: Comics[], total: number }> {
-        const response = await instance.get(`/comics`, {
-          params: {
-            apikey,
-            ts,
-            hash,
-            offset,
-          },
-        });
-
-        return { results: response.data.data.results, total: response.data.data.total };
-  },
-
-
-    async getComic(comicId: number): Promise<Comic> {
-        const response = await instance.get(`/comics/${comicId}`, {
-          params: {
-            apikey,
-            ts,
-            hash,
-          },
-        });
-
-        return response.data.data.results[0];
-  },
-
-    async searchCharactersByName(query: string, offset: number): Promise<{ results: Characters[], total: number }> {
+      async searchCharactersByName(query: string, offset: number): Promise<{ results: Characters[], total: number }> {
         const response = await instance.get(`/characters`, {
           params: {
             apikey,
@@ -72,18 +46,5 @@ export default {
         return { results: response.data.data.results, total: response.data.data.total };
   },
 
-    async searchComicsByTitle(query: string, offset: number): Promise<{ results: Comics[], total: number }> {
-        const response = await instance.get(`/comics`, {
-          params: {
-            apikey,
-            ts,
-            hash,
-            offset,
-            titleStartsWith: query,
-          },
-        });
-
-        return { results: response.data.data.results, total: response.data.data.total };
-  },
-
-};
+  };
+
