+++ react-ts-template/src/api/types/characters.ts
@@ -29,31 +29,3 @@ export interface ComicSummary{
     resurceURI: string;
     name: string;
 }
-
-export interface Comics {
-  id: number;
-  title: string;
-  description: string;
-  thumbnail: Image;
-}
-
-export interface Comic {
-  id: number;
-  title: string;
-  description: string;
-  thumbnail: Image;
-  characters: CharacterList;
-}
-
-export interface CharacterList {
-    available: number;
-    returned: number;
-    collectionURI: string;
-    items: CharacterSummary[];
-}
-
-export interface CharacterSummary {
-    resourceURI: string;
-    name: string;
-    role: string;
-}
